refactor(observables): rename files to match statics
