+++ src/Rx.KitchenSink.ts
@@ -22,20 +22,20 @@ export interface KitchenSinkOperators<T> extends CoreOperators<T> {
 import './operators/combineLatest-static';
 import './operators/concat-static';
 import './operators/merge-static';
-import './observables/DeferObservable';
-import './observables/EmptyObservable';
-import './observables/ForkJoinObservable';
-import './observables/FromObservable';
-import './observables/ArrayObservable';
-import './observables/FromEventObservable';
-import './observables/FromEventPatternObservable';
-import './observables/PromiseObservable';
-import './observables/CallbackObservable';
-import './observables/IntervalObservable';
-import './observables/InfiniteObservable';
-import './observables/RangeObservable';
-import './observables/ErrorObservable';
-import './observables/TimerObservable';
+import './observable/defer';
+import './observable/empty';
+import './observable/forkJoin';
+import './observable/from';
+import './observable/fromArray';
+import './observable/fromEvent';
+import './observable/fromEventPattern';
+import './observable/fromPromise';
+import './observable/fromCallback';
+import './observable/interval';
+import './observable/never';
+import './observable/range';
+import './observable/throw';
+import './observable/timer';
 import './operators/zip-static';
 
 // Operators
@@ -130,7 +130,7 @@ import {Subscription} from './Subscription';
 import {Subscriber} from './Subscriber';
 import {ReplaySubject} from './subjects/ReplaySubject';
 import {BehaviorSubject} from './subjects/BehaviorSubject';
-import {ConnectableObservable} from './observables/ConnectableObservable';
+import {ConnectableObservable} from './observable/ConnectableObservable';
 import {Notification} from './Notification';
 import {EmptyError} from './util/EmptyError';
 import {ArgumentOutOfRangeError} from './util/ArgumentOutOfRangeError';
