+++ src/Rx.ts
@@ -5,20 +5,20 @@ import {Observable} from './Observable';
 import './operators/combineLatest-static';
 import './operators/concat-static';
 import './operators/merge-static';
-import './observables/DeferObservable';
-import './observables/EmptyObservable';
-import './observables/ForkJoinObservable';
-import './observables/FromObservable';
-import './observables/ArrayObservable';
-import './observables/FromEventObservable';
-import './observables/FromEventPatternObservable';
-import './observables/PromiseObservable';
-import './observables/CallbackObservable';
-import './observables/IntervalObservable';
-import './observables/InfiniteObservable';
-import './observables/RangeObservable';
-import './observables/ErrorObservable';
-import './observables/TimerObservable';
+import './observable/defer';
+import './observable/empty';
+import './observable/forkJoin';
+import './observable/from';
+import './observable/fromArray';
+import './observable/fromEvent';
+import './observable/fromEventPattern';
+import './observable/fromPromise';
+import './observable/fromCallback';
+import './observable/interval';
+import './observable/never';
+import './observable/range';
+import './observable/throw';
+import './observable/timer';
 import './operators/zip-static';
 
 //operators
@@ -106,7 +106,7 @@ import {Subscription} from './Subscription';
 import {Subscriber} from './Subscriber';
 import {ReplaySubject} from './subjects/ReplaySubject';
 import {BehaviorSubject} from './subjects/BehaviorSubject';
-import {ConnectableObservable} from './observables/ConnectableObservable';
+import {ConnectableObservable} from './observable/ConnectableObservable';
 import {Notification} from './Notification';
 import {EmptyError} from './util/EmptyError';
 import {ArgumentOutOfRangeError} from './util/ArgumentOutOfRangeError';
