+++ src/observable/empty.ts
