+++ src/observable/fromEventPattern.ts
