+++ src/observable/never.ts
