+++ src/operators/every.ts
@@ -1,9 +1,9 @@
 import {Operator} from '../Operator';
 import {Observer} from '../Observer';
 import {Observable} from '../Observable';
-import {ScalarObservable} from '../observables/ScalarObservable';
-import {ArrayObservable} from '../observables/ArrayObservable';
-import {ErrorObservable} from '../observables/ErrorObservable';
+import {ScalarObservable} from '../observable/ScalarObservable';
+import {ArrayObservable} from '../observable/fromArray';
+import {ErrorObservable} from '../observable/throw';
 import {Subscriber} from '../Subscriber';
 import {tryCatch} from '../util/tryCatch';
 import {errorObject} from '../util/errorObject';
